HW6: small refactoring
