--- src/actions/index.js
@@ -1,7 +1,0 @@
-const UPDATE_MOVIE_ACTION = {
-    type: 'UPDATE_MOVIE',
-    payload: {
-        // title
-        //
-    }
-}
